+++ pages/blocks/authors/authors.js
@@ -9,8 +9,7 @@
  * OF ANY KIND, either express or implied. See the License for the specific language
  * governing permissions and limitations under the License.
  */
-import { getMetadata, fetchPlaceholders } from '../../scripts/scripts.js';
-import { createTag } from '../../scripts/utils.js';
+import { getMetadata, createTag, fetchPlaceholders } from '../../scripts/utils.js';
 
 export default async function decorate(block) {
   const metaAuthors = getMetadata('authors');
