+++ pages/blocks/course/course.css
@@ -47,6 +47,7 @@ main .course .video-player-list {
   padding: 24px;
   padding-left: 48px;
   background-color: #f4f4f4;
+  margin: 0;
 }
 
 main .course .video-player-list-link {
