+++ pages/scripts/scripts.js
@@ -18,7 +18,6 @@ import {
   customSpacings,
   externalLinks,
   gnavUnderline,
-  toClassName,
 } from './utils.js';
 
 const LIBS = 'https://milo.adobe.com/libs';
@@ -76,44 +75,3 @@ const { loadArea, loadDelayed, setConfig } = await import(`${miloLibs}/utils/uti
   loadDelayed();
 }());
 
-export function getLocale(url) {
-  const locale = url.pathname.split('/')[1];
-  if (/^[a-z]{2}$/.test(locale)) {
-    return locale;
-  }
-  return 'pages';
-}
-
-/**
- * Retrieves the content of a metadata tag.
- * @param {string} name The metadata name (or property)
- * @returns {string} The metadata value
- */
-export function getMetadata(name) {
-  const attr = name && name.includes(':') ? 'property' : 'name';
-  const meta = [...document.head.querySelectorAll(`meta[${attr}="${name}"]`)].map((el) => el.content).join(', ');
-  return meta;
-}
-
-export async function fetchPlaceholders() {
-  if (!window.placeholders) {
-    try {
-      const locale = getLocale(window.location);
-      const urlPrefix = locale === 'us' ? '' : `/${locale}`;
-      const resp = await fetch(`${urlPrefix}/pages/artisthub/placeholders.json`);
-      const json = await resp.json();
-      window.placeholders = {};
-      json.data.forEach((placeholder) => {
-        window.placeholders[toClassName(placeholder.Key)] = placeholder.Text;
-      });
-    } catch {
-      const resp = await fetch('/pages/artisthub/placeholders.json');
-      const json = await resp.json();
-      window.placeholders = {};
-      json.data.forEach((placeholder) => {
-        window.placeholders[toClassName(placeholder.Key)] = placeholder.Text;
-      });
-    }
-  }
-  return window.placeholders;
-}
