+++ pages/templates/artisthub/artisthub.css
@@ -14,49 +14,6 @@
   @media screen and (min-width: 1120px)
 */
 
-/* customize gnav for artisthub */
-
-body nav.gnav .gnav-brand {
-  justify-content: flex-start
-}
-
-body nav.gnav .gnav-brand picture {
-  width: 100%;
-  height: 100%;
-  display: flex;
-  align-items: center;
-  justify-content: flex-start;
-}
-
-body nav.gnav .gnav-brand picture img {
-  width: auto;
-  height: 100%;
-}
-
-@media (min-width: 1200px) {
-  .gnav-navitem.has-Menu.is-Open .gnav-navitem-menu li,
-  .gnav-navitem.has-Menu.is-Open .gnav-navitem-menu ul {
-    padding: 0;
-  }
-  .gnav-navitem.has-Menu.is-Open .gnav-navitem-menu ul li a {
-    padding-top: 8px;
-    padding-bottom: 8px;
-  }
-}
-
-body header.gnav {
-  position: fixed;
-  top: 0;
-  left: 0;
-  width: 100%;
-  z-index: 1000;
-}
-
-body header.gnav + main {
-  margin-top: 65px;
-}
-
-
 /* cusomize typography for artisthub */
 
 p {
